+++ tgbotapi.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/expectations/WaitDeepLinks.kt
@@ -1,0 +1,23 @@
+package dev.inmo.tgbotapi.extensions.behaviour_builder.expectations
+
+import dev.inmo.tgbotapi.extensions.behaviour_builder.BehaviourContext
+import dev.inmo.tgbotapi.extensions.utils.regularTextSourceOrNull
+import dev.inmo.tgbotapi.requests.abstracts.Request
+import dev.inmo.tgbotapi.types.message.abstracts.CommonMessage
+import dev.inmo.tgbotapi.types.message.content.TextContent
+import dev.inmo.tgbotapi.types.message.textsources.RegularTextSource
+import kotlinx.coroutines.flow.*
+
+suspend fun BehaviourContext.waitDeepLinks(
+    initRequest: Request<*>? = null,
+    errorFactory: NullableRequestBuilder<*> = { null },
+): Flow<Pair<CommonMessage<TextContent>, String>> = waitCommandMessage(
+    "start",
+    initRequest,
+    errorFactory
+)
+    .requireSingleCommand()
+    .requireCommandAtStart()
+    .flattenCommandsWithParams().mapNotNull {
+        it.first to (it.second.second.singleOrNull() ?.regularTextSourceOrNull() ?.source ?.removePrefix(" ") ?: return@mapNotNull null)
+    }