+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # TelegramBotAPI changelog
 
+## 3.2.0
+
+**Since this update, `RequestsExecutor#execute` may throw only `BotException`. In case you wish to handle some exceptions from `execute` you must catch `BotException` and handle its `cause`**
+
+* `Versions`:
+    * `Serialization`: `1.4.0-RC` -> `1.4.0`
+    * `MicroUtils`: `0.12.1` -> `0.12.4`
+* `Core`:
+  * `SetWebhook#allowedUpdates` now is `ALL_UPDATES_LIST` by default instead of `null`
+* `API`:
+  * Extension `TelegramBot#setWebhook` parameter `allowedUpdates` now is `ALL_UPDATES_LIST` by default instead of `null`
+* `Utils`:
+  * All related to long polling extensions parameters `allowedUpdates` now are `ALL_UPDATES_LIST` by default instead of `null`
+
 ## 3.1.1
 
 * `Common`: